refactor: move org endpoints to dedicated api file
